Remote
A bot designed to allow me to remotely control my laptop using the discord mobile app in conjunction with chrome remote desktop.

The core classes are identical to FFBEBot and only the special classes will be mentioned

It implements some of the utility classes macro and shellutil used to send command line commands and execute custom macros

The commands deal with controlling the computer through the use of java's Robot class. In addition it can execute any shell command with its shell classes and execute commands delayed after a certain period.